requirements.txt
python-dotenv
openai
pyupbit
ta
google-api-python-client
youtube_transcript_api